requirements.txt
librosa
numpy
pandas
scikit-learn
pyserial
matplotlib
